+++ RenderSystems/GLES2/src/OgreGLES2PixelFormat.cpp
@@ -982,10 +982,10 @@ namespace Ogre {
         };
     }
     //-----------------------------------------------------------------------------
-    size_t GLES2PixelUtil::getMaxMipmaps(uint32 width, uint32 height, uint32 depth,
+    uint32 GLES2PixelUtil::getMaxMipmaps(uint32 width, uint32 height, uint32 depth,
                                       PixelFormat format)
     {
-        size_t count = 0;
+        uint32 count = 0;
         if((width > 0) && (height > 0) && (depth > 0))
         {
             do {
